+++ src/core_modules/capture-core/components/DataEntries/EnrollmentRegistrationEntry/EnrollmentRegistrationEntry.component.js
@@ -7,7 +7,7 @@ import { compose } from 'redux';
 import { useHistory } from 'react-router-dom';
 import { useScopeInfo } from '../../../hooks/useScopeInfo';
 import { scopeTypes } from '../../../metaData';
-import { ConfirmDialog } from '../../Dialogs/ConfirmDialog.component';
+import { DiscardDialog } from '../../Dialogs/DiscardDialog.component';
 import { EnrollmentDataEntry } from '../Enrollment';
 import type { Props, PlainProps } from './EnrollmentRegistrationEntry.types';
 import { withSaveHandler } from '../../DataEntry';
@@ -16,7 +16,7 @@ import { InfoIconText } from '../../InfoIconText';
 import { withErrorMessagePostProcessor } from '../withErrorMessagePostProcessor';
 import { buildUrlQueryString } from '../../../utils/routing';
 import { withDuplicateCheckOnSave } from '../common/TEIAndEnrollment/DuplicateCheckOnSave';
-import { defaultDialogProps } from '../../Dialogs/ConfirmDialog.constants';
+import { defaultDialogProps } from '../../Dialogs/DiscardDialog.constants';
 
 const styles = () => ({
     actions: {
@@ -128,9 +128,9 @@ const EnrollmentRegistrationEntryPlain =
                       </InfoIconText>
                   </>
               }
-              <ConfirmDialog
+              <DiscardDialog
                   {...defaultDialogProps}
-                  onConfirm={navigateToWorkingListsPage}
+                  onDestroy={navigateToWorkingListsPage}
                   open={!!showWarning}
                   onCancel={() => { setShowWarning(false); }}
               />